// escrow/ui/app/src/contracts/EscrowContractAbi.d.ts
export type IdentityInput = Enum<{
  Address: AddressInput;
  ContractId: ContractIdInput;
}>;
// ...
export type AddressInput = { value: string };
export type AddressOutput = AddressInput;
export type ArbiterInput = {
  address: IdentityInput;
  asset: ContractIdInput;
  fee_amount: BigNumberish;
};
export type ArbiterOutput = {
  address: IdentityOutput;
  asset: ContractIdOutput;
  fee_amount: BN;
};
export type AssetInput = { amount: BigNumberish; id: ContractIdInput };
export type AssetOutput = { amount: BN; id: ContractIdOutput };
export type BuyerInput = {
  address: IdentityInput;
  asset: Option<ContractIdInput>;
  deposited_amount: BigNumberish;
};
export type BuyerOutput = {
  address: IdentityOutput;
  asset: Option<ContractIdOutput>;
  deposited_amount: BN;
};
export type ContractIdInput = { value: string };